acc test fixes 2 back button
