+++ web/src/components/organisms/Asset/Asset/index.tsx
@@ -1,13 +1,17 @@
 import styled from "@emotion/styled";
-import { useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 
+import Button from "@reearth-cms/components/atoms/Button";
+import PageHeader from "@reearth-cms/components/atoms/PageHeader";
 import AssetBody from "@reearth-cms/components/molecules/Asset/Asset/AssetBody";
-import AssetHeader from "@reearth-cms/components/molecules/Asset/Asset/AssetHeader";
+import { useT } from "@reearth-cms/i18n";
 
 import useHooks from "./hooks";
 
 const Asset: React.FC = () => {
-  const { assetId } = useParams();
+  const t = useT();
+  const navigate = useNavigate();
+  const { workspaceId, projectId, assetId } = useParams();
   const {
     asset,
     updateAsset,
@@ -25,12 +29,20 @@ const Asset: React.FC = () => {
     }
   };
 
+  const handleBack = () => {
+    navigate(`/workspace/${workspaceId}/project/${projectId}/asset/`);
+  };
+
   return isLoading ? (
     // TODO: need to add a spinner
     <Wrapper>loading...</Wrapper>
   ) : asset ? (
     <Wrapper>
-      <AssetHeader title={`Asset/${asset?.fileName}`} handleSave={handleSave} />
+      <PageHeader
+        title={`Asset/${asset?.fileName}`}
+        onBack={handleBack}
+        extra={<Button onClick={handleSave}>{t("Save")}</Button>}
+      />
       <AssetBody
         asset={asset}
         selectedPreviewType={selectedPreviewType}
