+++ web/src/components/organisms/Asset/AssetList/index.tsx
@@ -1,9 +1,12 @@
 import styled from "@emotion/styled";
 import { useParams } from "react-router-dom";
 
+import PageHeader from "@reearth-cms/components/atoms/PageHeader";
+import { UploadProps } from "@reearth-cms/components/atoms/Upload";
 import { Asset } from "@reearth-cms/components/molecules/Asset/asset.type";
-import AssetListHeader from "@reearth-cms/components/molecules/Asset/AssetList/AssetListHeader";
 import AssetListTable from "@reearth-cms/components/molecules/Asset/AssetList/AssetListTable";
+import UploadAsset from "@reearth-cms/components/molecules/Asset/UploadAsset";
+import { fileFormats, imageFormats } from "@reearth-cms/components/molecules/Common/Asset";
 
 import useHooks from "./hooks";
 
@@ -50,17 +53,41 @@ const AssetList: React.FC = () => {
     navigate(`/workspace/${workspaceId}/project/${projectId}/asset/${asset.id}`);
   };
 
+  const uploadProps: UploadProps = {
+    name: "file",
+    multiple: true,
+    directory: false,
+    showUploadList: true,
+    accept: imageFormats + "," + fileFormats,
+    listType: "picture",
+    onRemove: file => {
+      const index = fileList.indexOf(file);
+      const newFileList = fileList.slice();
+      newFileList.splice(index, 1);
+      setFileList(newFileList);
+    },
+    beforeUpload: (_file, files) => {
+      setFileList([...fileList, ...files]);
+      return false;
+    },
+    fileList,
+  };
+
   return (
     <Wrapper>
-      <AssetListHeader
+      <PageHeader
         title="Asset"
-        fileList={fileList}
-        uploading={uploading}
-        uploadModalVisibility={uploadModalVisibility}
-        displayUploadModal={displayUploadModal}
-        hideUploadModal={hideUploadModal}
-        handleUpload={handleUpload}
-        setFileList={setFileList}
+        extra={
+          <UploadAsset
+            fileList={fileList}
+            uploading={uploading}
+            uploadProps={uploadProps}
+            uploadModalVisibility={uploadModalVisibility}
+            displayUploadModal={displayUploadModal}
+            hideUploadModal={hideUploadModal}
+            handleUpload={handleUpload}
+          />
+        }
       />
       <AssetListTable
         assetList={assetList}
